pyproject.toml
[project]
name = "ascii-art-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "NAGATA Minoru", email = "nagata@lean-system.co.jp" }
]
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.49.1",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
]

[project.scripts]
ascii-art-maker = "ascii_art_maker:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
